requirements.txt
# Core ML/AI libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Data processing and analysis
scipy>=1.10.0
statsmodels>=0.14.0
ta-lib>=0.4.0
yfinance>=0.2.0

# Communication and networking
pyzmq>=25.0.0
websockets>=11.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Monitoring and visualization
dash>=2.14.0
plotly>=5.15.0
dash-bootstrap-components>=1.4.0
flask>=2.3.0

# Database and storage
sqlite3
sqlalchemy>=2.0.0
redis>=4.6.0
pymongo>=4.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
colorama>=0.4.6
tqdm>=4.65.0
joblib>=1.3.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Performance and optimization
numba>=0.57.0
cython>=3.0.0
psutil>=5.9.0

# Financial data and analysis
ccxt>=4.0.0
alpha-vantage>=2.3.0
quandl>=3.7.0

# Risk management
arch>=6.2.0  # GARCH models
pyfolio>=0.9.2  # Portfolio analysis

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0
grafana-api>=1.0.3

# Async and concurrency
asyncio
concurrent.futures
threading
multiprocessing

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0